requirements.txt
asgiref==3.5.0
colorama==0.4.4
Django==4.0.3
django-crispy-forms==1.14.0
Pillow==9.0.1
qrcode==7.3.1
sqlparse==0.4.2
tzdata==2022.1
